CPG v1.0
Forgot to upload before, but this is what I had done 2 presentations ago. For the next step, I need to modify the code to take in an array as an input rather than lines from the command input and output an array. I still might keep the print function, or I may make an hfile with the generator and the print function, so the print function can be called separately.